+++ src/mulle-utf/mulle-utf16-string.h
@@ -21,16 +21,16 @@
 
 
 MULLE__UTF_GLOBAL
-unsigned int  mulle_utf16_strnlen( mulle_utf16_t *src, unsigned int len);
+size_t  mulle_utf16_strnlen( mulle_utf16_t *src, size_t len);
 
 MULLE__UTF_GLOBAL
 mulle_utf16_t  *mulle_utf16_strdup( mulle_utf16_t *s);
 
 MULLE__UTF_GLOBAL
-mulle_utf16_t  *mulle_utf16_strncpy( mulle_utf16_t *dst, unsigned int len, mulle_utf16_t *src);
+mulle_utf16_t  *mulle_utf16_strncpy( mulle_utf16_t *dst, size_t len, mulle_utf16_t *src);
 
 MULLE__UTF_GLOBAL
-int             mulle_utf16_strncmp( mulle_utf16_t *s1, mulle_utf16_t *s2, unsigned int len);
+int             mulle_utf16_strncmp( mulle_utf16_t *s1, mulle_utf16_t *s2, size_t len);
 
 MULLE__UTF_GLOBAL
 mulle_utf16_t  *mulle_utf16_strchr( mulle_utf16_t *s, mulle_utf32_t c);  // sic
@@ -42,10 +42,10 @@ MULLE__UTF_GLOBAL
 mulle_utf16_t  *mulle_utf16_strstr( mulle_utf16_t *s1, mulle_utf16_t *s2);
 
 MULLE__UTF_GLOBAL
-unsigned int    mulle_utf16_strspn( mulle_utf16_t *s1, mulle_utf16_t *s2);
+size_t    mulle_utf16_strspn( mulle_utf16_t *s1, mulle_utf16_t *s2);
 
 MULLE__UTF_GLOBAL
-unsigned int    mulle_utf16_strcspn( mulle_utf16_t *s1, mulle_utf16_t *s2);
+size_t    mulle_utf16_strcspn( mulle_utf16_t *s1, mulle_utf16_t *s2);
 
 
 static inline int   mulle_utf16_strcmp( mulle_utf16_t *s1, mulle_utf16_t *s2)
@@ -62,13 +62,13 @@ static inline int   mulle_utf16_atoi( mulle_utf16_t *s)
 
 
 
-static inline void   mulle_utf16_memcpy( mulle_utf16_t *dst, mulle_utf16_t *src, unsigned int len)
+static inline void   mulle_utf16_memcpy( mulle_utf16_t *dst, mulle_utf16_t *src, size_t len)
 {
    memcpy( dst, src, sizeof( mulle_utf16_t) * len);
 }
 
 
-static inline void   mulle_utf16_memmove( mulle_utf16_t *dst, mulle_utf16_t *src, unsigned int len)
+static inline void   mulle_utf16_memmove( mulle_utf16_t *dst, mulle_utf16_t *src, size_t len)
 {
    memmove( dst, src, sizeof( mulle_utf16_t) * len);
 }
