+++ src/mulle-utf/mulle-utf16.c
@@ -82,14 +82,14 @@ int  mulle_utf16_is_valid_surrogatepair( mulle_utf16_t hi, mulle_utf16_t lo)
  * copy converters
  */
 mulle_utf32_t  *_mulle_utf16_convert_to_utf32( mulle_utf16_t *src,
-                                               unsigned int len,
+                                               size_t len,
                                                mulle_utf32_t *dst)
 {
    mulle_utf16_t   *sentinel;
    mulle_utf32_t   x;
 
    // if dst_len == -1, then sentinel - 1 = dst_sentinel (OK!)
-   assert( len != (unsigned int) -1);
+   assert( len != (size_t) -1);
 
    sentinel = &src[ len];
 
@@ -112,14 +112,14 @@ mulle_utf32_t  *_mulle_utf16_convert_to_utf32( mulle_utf16_t *src,
 //
 // this also does not do any error checking,
 // must be proper UTF16 code!
-char  *_mulle_utf16_convert_to_utf8( mulle_utf16_t *src, unsigned int len, char *_dst)
+char  *_mulle_utf16_convert_to_utf8( mulle_utf16_t *src, size_t len, char *_dst)
 {
    unsigned char   *dst = (unsigned char *) _dst;
    mulle_utf16_t   *sentinel;
    mulle_utf32_t   x;
 
    // if dst_len == -1, then sentinel - 1 = dst_sentinel (OK!)
-   assert( len != (unsigned int) -1);
+   assert( len != (size_t) -1);
 
    sentinel = &src[ len];
 
@@ -181,7 +181,7 @@ recheck:
 // this also does not do any error checking,
 // must be proper UTF16 code!
 void  mulle_utf16_bufferconvert_to_utf8( mulle_utf16_t *src,
-                                         unsigned int len,
+                                         size_t len,
                                          void *buffer,
                                          mulle_utf_add_bytes_function_t *addbytes)
 {
@@ -191,7 +191,7 @@ void  mulle_utf16_bufferconvert_to_utf8( mulle_utf16_t *src,
    unsigned char   *s_flush;
    unsigned char   tmp[ 128];
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
 
    // if dst_len == -1, then sentinel - 1 = dst_sentinel (OK!)
@@ -261,7 +261,7 @@ recheck:
 
 
 void  mulle_utf16_bufferconvert_to_utf32( mulle_utf16_t *src,
-                                          unsigned int len,
+                                          size_t len,
                                           void *buffer,
                                           mulle_utf_add_bytes_function_t *addbytes)
 {
@@ -271,7 +271,7 @@ void  mulle_utf16_bufferconvert_to_utf32( mulle_utf16_t *src,
    mulle_utf32_t   *s_flush;
    mulle_utf32_t   tmp[ 32];
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
 
    // if dst_len == -1, then sentinel - 1 = dst_sentinel (OK!)
@@ -313,7 +313,7 @@ static inline int  mulle_utf16_is_invalid_char( mulle_utf16_t c)
 //
 // just checks that the surrogate pairs are ok
 //
-mulle_utf16_t  *mulle_utf16_validate( mulle_utf16_t *src, unsigned int len)
+mulle_utf16_t  *mulle_utf16_validate( mulle_utf16_t *src, size_t len)
 {
    mulle_utf16_t   c;
    mulle_utf16_t   d;
@@ -322,7 +322,7 @@ mulle_utf16_t  *mulle_utf16_validate( mulle_utf16_t *src, unsigned int len)
    if( ! src)
       return( NULL);
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
 
    sentinel = &src[ len];
@@ -362,13 +362,13 @@ mulle_utf16_t  *mulle_utf16_validate( mulle_utf16_t *src, unsigned int len)
 //
 // this routine does not validate...
 //
-size_t  mulle_utf16_utf8length( mulle_utf16_t *src, unsigned int len)
+size_t  mulle_utf16_utf8length( mulle_utf16_t *src, size_t len)
 {
    mulle_utf16_t   c;
    mulle_utf16_t   *sentinel;
    size_t          size;
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
    if( ! len)
       return( 0);
@@ -405,12 +405,12 @@ size_t  mulle_utf16_utf8length( mulle_utf16_t *src, unsigned int len)
 }
 
 
-unsigned int  mulle_utf16_utf32length( mulle_utf16_t *src, unsigned int len)
+size_t  mulle_utf16_utf32length( mulle_utf16_t *src, size_t len)
 {
    mulle_utf16_t   c;
    mulle_utf16_t   *sentinel;
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
    if( ! len)
       return( 0);
@@ -434,13 +434,13 @@ unsigned int  mulle_utf16_utf32length( mulle_utf16_t *src, unsigned int len)
 
 
 
-unsigned int  mulle_utf16_length( mulle_utf16_t *src, unsigned int len)
+size_t  mulle_utf16_length( mulle_utf16_t *src, size_t len)
 {
    mulle_utf16_t   c;
    mulle_utf16_t   *sentinel;
-   unsigned int          dst_len;
+   size_t          dst_len;
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
 
    sentinel = &src[ len];
@@ -513,7 +513,7 @@ mulle_utf32_t   _mulle_utf16_previous_utf32character( mulle_utf16_t **s_p)
 // a long or long long
 // (b) masking value with 0x80808080 to figure out if all are "ASCII"
 //
-int  mulle_utf16_information( mulle_utf16_t *src, unsigned int len, struct mulle_utf_information *info)
+int  mulle_utf16_information( mulle_utf16_t *src, size_t len, struct mulle_utf_information *info)
 {
    mulle_utf16_t                  _c;
    mulle_utf16_t                  *start;
@@ -539,7 +539,7 @@ int  mulle_utf16_information( mulle_utf16_t *src, unsigned int len, struct mulle
    if( ! src)
       goto fail;
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( src);
 
    //
@@ -617,14 +617,14 @@ fail:
 
 
 int   mulle_utf16_contains_character_larger_or_equal( mulle_utf16_t *s,
-                                                      unsigned int len,
+                                                      size_t len,
                                                       mulle_utf16_t d)
 {
    mulle_utf16_t   _c;
    mulle_utf16_t   *sentinel;
    mulle_utf16_t   *p;
 
-   if( len == (unsigned int) -1)
+   if( len == (size_t) -1)
       len = mulle_utf16_strlen( s);
 
    p        = s;
@@ -643,7 +643,7 @@ int   mulle_utf16_contains_character_larger_or_equal( mulle_utf16_t *s,
 //
 // src must be known to be UTF15, and contain no zeroes
 //
-enum mulle_utf_charinfo   _mulle_utf16_charinfo( mulle_utf16_t *src, unsigned int len)
+enum mulle_utf_charinfo   _mulle_utf16_charinfo( mulle_utf16_t *src, size_t len)
 {
    mulle_utf16_t   _c;
    mulle_utf16_t   *start;
